[오혜인] 1) vo - MemberFile fileSize int-> long 수정

+++ lms/src/main/java/kr/co/gdu/lms/vo/MemberFile.java
@@ -9,6 +9,6 @@ public class MemberFile {
 	private String memberFileName;
 	private String memberFileOriginName;
 	private String memberFileType;
-	private int memberFileSize;
+	private long memberFileSize;
 	private String memberFileCreateDate;
 }
